convert calls to syntax.error into syntax.print_error
this allows better interactive use in emacs, because it shows all
errors instead of just the first one. however, it makes things quite
subtle, since the resolver can no longer assume that checks have
passed. there are definitely lots of cases where I didn't bother to
bail out correctly, so we will need to continue to iterate on this.