Fixed liveness bug in raft spec
The transition receive_append_msg had some predicates over the given
node's current log in the next state that should have been over the
node's current log in the old state. This manifested itself as a sort
of liveness bug, in that the only receive_append_msg transition that
could occur was the noop one (the two else branches). The trace sanity
check did not catch this because it produces a trace in which there is
only one node, and that node sends an append_reply_msg in
send_append_message. To make the trace sanity check more meaningful, I
removed the append_reply_msg from send_append_message, since the
algorithm works even if a leader doesn't reply to its own append_msg
in the same atomic step as sending it. This way, to commit, a real
receive_append_msg step has to occur.