fix name shadowing problem in z3 translation
mypyvy is carefully handles name shadowing, but we had assumed that
distinct calls to z3.Const returned distinct declarations. This is not
the case: z3 considers to different declarations to be equal if their
names are equal. We fix this by adding a nonce to names of constants
declared for bound variables.