first draft of what I'm calling "subclause completion of a diagram"
The idea is that we want the universal clause that is the negation of
the diagram to have the property that every stronger clause is a
subclause. For this to be true, we need to "complete" the diagram by
introducing fresh variables for every subexpression, and adding
equalities to the diagram relating the fresh variables to their
expression.

Thus, for example,
    exists x. r(x,x)
completes to
    exists x, y, z. r(y, z) & x = y & x = z & y = z

This is implemented using a union-find data structure to track
equivalence classes of variables. All pairs in an equivalence class
are asserted equal in the diagram, and all elements of all pairs of
distinct classes are asserted not equal. This leads to large diagrams
very quickly, and perhaps can be improved.

For now, this feature is hidden behind the flag
--diagrams-subclause-complete because:

It seems to cause both updr and pd-repeated-houdini to take longer,
especially pd. It's possible that the current completion method is
just too dumb and creates intractably large completed diagrams.