better treatment of undefined sorts in z3 models
the situation is that the z3 model defines a relation, but does not
define all the sorts in the relation's domain. the old
code (_old_model_to_trace) silently drops such relations, and the code
for model_to_first_order_structure before this commit resulted in a
runtime error in such situations. the new code keeps the relations and
adds a single element universe for undefined sorts. this actually
occurs in pd-primal-dual-houdini on stoppable paxos, e.g. on
examples/pd/stoppable_paxos_dnf_forall.pyv.