dual_houdini_frames: remove most calls to forward_explore_from_predicates
I suspect most calls were made thinking they would be cheap due to
caching, but that's not the case now, and I speculate that it would be
better not to call forward_explore_from_predicates so aggressively. In
any case, in a better find_dual_edge with BFS over states and number
of ps, we would get the effect of forward_explore_from_predicates.